[feature] Get registration-by-script to work
As documented (in a somewhat sybilline way) in the [2.3 release
notes](https://www.theforeman.org/manuals/2.3/index.html), there now
exists a register-by-script feature: a node that is already live (with
distro installed) can enroll into Foreman by running some `curl |
bash` script. The things to do as part of this script are computed by
Foreman on the fly.

- (At least for the time being) we need a “real” Foreman CA for this
to work. Make Træfik's SSL cert visible to the Foreman frontend
- Configure `ssl_ca_file` setting to point to same
- Add `Linux registration default` as a provisioning template of kind
`registration` to Ubuntu 20.04 OS (this is required for some
second-stage script that is being curl | bash'd out of the first one)