[feature] kexec directly from the “host details” screen
We really don't want to have to delete a host in order to reinstall
it (as we currently must under the “discovery → auto provision”
workflow) - This would cause all the precious host metadata, such
as its `etcd_member_name` etc. to be lost. Instead, make it so the
“Build” button on the “host details” screen does a kexec when
possible.

On hosts that are currently running the Foreman discovery image (and
thus respond to the `/facts` and `/kexec` APIs on port 8443):
- Hijack the existing UI and control flows that Foreman provides
out-of-the-box for hosts whose power can be controlled by the smart
proxy
- Hide all errors
- When the “reboot now” checkbox is ticked, perform a kexec instead of
a reboot

The “how”:
- Do all of the above using an `ActiveSupport::Concern` hijack installed
at Rails `config.to_prepare` time (like I saw it done in the foreman_discovery
plugin for the very same `HostsController` class)
- Remove the discovery rule management from the configuration-as-code,
since we will only be provisioning using the new method from now on